--- a/TaskOrchestrator.sln
+++ b/TaskOrchestrator.sln
@@ -12,6 +12,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Misc", "Misc", "{8EC462FD-D
 		.gitattributes = .gitattributes
 		.gitignore = .gitignore
 		LICENSE.txt = LICENSE.txt
+		Patch_Notes_1.0.1.md = Patch_Notes_1.0.1.md
 		README.md = README.md
 	EndProjectSection
 EndProject
@@ -21,6 +22,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{
 	ProjectSection(SolutionItems) = preProject
 		.github\workflows\cicd.yml = .github\workflows\cicd.yml
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Benchmarks", "Benchmarks\Benchmarks.csproj", "{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,6 +59,18 @@ Global
 		{A2085A7B-7E41-469C-A481-9420509D256C}.Release|x64.Build.0 = Release|Any CPU
 		{A2085A7B-7E41-469C-A481-9420509D256C}.Release|x86.ActiveCfg = Release|Any CPU
 		{A2085A7B-7E41-469C-A481-9420509D256C}.Release|x86.Build.0 = Release|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Debug|x64.Build.0 = Debug|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Debug|x86.Build.0 = Debug|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Release|x64.ActiveCfg = Release|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Release|x64.Build.0 = Release|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Release|x86.ActiveCfg = Release|Any CPU
+		{F1E3BA3D-8035-4D44-9304-3B3E2943B16C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE