--- a/TaskOrchestrator/Core/AsyncPool.cs
+++ b/TaskOrchestrator/Core/AsyncPool.cs
@@ -1,4 +1,5 @@
-﻿using TaskOrchestrator.Internal;
+﻿using System.Collections.Concurrent;
+using TaskOrchestrator.Internal;
 using TaskOrchestrator.Utils;
 
 namespace TaskOrchestrator.Core;
@@ -12,10 +13,10 @@ internal class AsyncPool : IDisposable
 
     private readonly CancellationTokenSource _cts = new();
     private readonly List<Task> _workers = new();
-    private readonly List<Task> _elasticWorkers = new();
+    private readonly ConcurrentBag<Task> _elasticWorkers = new();
+    private volatile int _elasticWorkersCount = 0;
     private readonly SemaphoreSlim _signal;
     private readonly AsyncQueue _queue = new();
-    private readonly object _elasticLock = new();
 
     public AsyncPool(Options options)
     {
@@ -31,22 +32,18 @@ internal class AsyncPool : IDisposable
 
     public void Enqueue(Func<Task> asyncAction, int weight = 0)
     {
-        lock (_elasticLock)
-        {
-            _queue.Enqueue(new Identified<Func<Task>>(asyncAction, weight), weight);
-            _signal.Release();
-            if (_signal.CurrentCount == 0)
-            {
-                EnsureElasticWorkers();
-            }
-        }
+        _queue.Enqueue(asyncAction, weight);
+        _signal.Release();
+        
+        if (_signal.CurrentCount == 0)
+            EnsureElasticWorkers();
     }
 
     public int PendingWorkCount()
         => _queue.PendingWorkCount;
 
     public int CurrentElasticWorkersCount()
-        => _elasticWorkers.Count;
+        => _elasticWorkersCount;
 
     public void Dispose()
     {
@@ -67,16 +64,18 @@ internal class AsyncPool : IDisposable
     #region PRIVATE
     private void EnsureElasticWorkers()
     {
-        if (_queue.IsEmpty || _elasticWorkers.Count >= Options.MaxElasticWorkers)
+        if (_queue.IsEmpty || _elasticWorkersCount >= Options.MaxElasticWorkers)
             return;
 
-        lock (_elasticLock)
+        if (Interlocked.CompareExchange(ref _elasticWorkersCount, 
+            _elasticWorkersCount + 1, _elasticWorkersCount) < Options.MaxElasticWorkers)
         {
-            if (_elasticWorkers.Count < Options.MaxElasticWorkers && !_queue.IsEmpty)
-            {
-                var workerTask = Task.Run(() => ElasticWorkerLoopAsync(_cts.Token));
-                _elasticWorkers.Add(workerTask);
-            }
+            var workerTask = Task.Run(() => ElasticWorkerLoopAsync(_cts.Token));
+            _elasticWorkers.Add(workerTask);
+        }
+        else
+        {
+            Interlocked.Decrement(ref _elasticWorkersCount);
         }
     }
 
@@ -95,52 +94,47 @@ internal class AsyncPool : IDisposable
 
             if (_queue.TryDequeue(out var task))
             {
-                if (task != null)
-                    await ExecuteTaskAsync(task);
+                await ExecuteTaskAsync(task);
             }
         }
     }
 
     private async Task ElasticWorkerLoopAsync(CancellationToken token)
     {
-        int? taskId = Task.CurrentId;
-
-        while (!token.IsCancellationRequested)
+        try
         {
-            try
+            while (!token.IsCancellationRequested)
             {
-                await _signal.WaitAsync(Options.ElasticWorkersTimeout, token);
-            }
-            catch (OperationCanceledException)
-            {
-                break;
-            }
-            catch (TimeoutException)
-            {
-                break;
-            }
+                try
+                {
+                    await _signal.WaitAsync(Options.ElasticWorkersTimeout, token);
+                }
+                catch (OperationCanceledException)
+                {
+                    break;
+                }
+                catch (TimeoutException)
+                {
+                    break;
+                }
 
-            if (_queue.TryDequeue(out var task))
-            {
-                if (task != null)
+                if (_queue.TryDequeue(out var task))
+                {
                     await ExecuteTaskAsync(task);
-            }
-            else
-            {
-                break;
+                }
+                else
+                {
+                    break;
+                }
             }
         }
-
-        lock (_elasticLock)
+        finally
         {
-            if (taskId.HasValue)
-            {
-                _elasticWorkers.RemoveAll(t => t.Id == taskId.Value);
-            }
+            Interlocked.Decrement(ref _elasticWorkersCount);
         }
     }
 
-    private async Task ExecuteTaskAsync(Identified<Func<Task>> task)
+    private async Task ExecuteTaskAsync(TaskItem<Func<Task>> task)
     {
         try
         {