--- a/TaskOrchestrator/Internal/AsyncQueue.cs
+++ b/TaskOrchestrator/Internal/AsyncQueue.cs
@@ -4,17 +4,18 @@ namespace TaskOrchestrator.Internal;
 
 internal class AsyncQueue
 {
-    private readonly PriorityQueue<Identified<Func<Task>>, int> _asyncQueue = new();
+    private readonly PriorityQueue<TaskItem<Func<Task>>, int> _asyncQueue = new();
     private readonly object _lock = new();
 
-    public void Enqueue(Identified<Func<Task>> item, int priority)
+    public void Enqueue(Func<Task> entity, int weight)
     {
         lock (_lock)
         {
-            _asyncQueue.Enqueue(item, -priority);
+            _asyncQueue.Enqueue(new TaskItem<Func<Task>>(entity, weight), -weight);
         }
     }
-    public bool TryDequeue(out Identified<Func<Task>>? item)
+
+    public bool TryDequeue(out TaskItem<Func<Task>> item)
     {
         lock (_lock)
         {
@@ -27,6 +28,7 @@ internal class AsyncQueue
             return false;
         }
     }
+
     public bool IsEmpty
     {
         get { lock (_lock) return _asyncQueue.Count == 0; }
@@ -34,6 +36,6 @@ internal class AsyncQueue
 
     public int PendingWorkCount
     {
-        get { return _asyncQueue.Count; }
+        get { lock (_lock) return _asyncQueue.Count; }
     }
 }